rename: 컴포넌트 구조 변경으로 인한 경로 이동 (#11)
